Adapt Codpiece maximization to combined modifiers
Support the All Attributes and All Attributes Percent objectives added
on main. Bound every combined modifier conservatively when component
values cross zero, while preserving explicit nonzero combined values.

Avoid using negative-objective contribution bounds to reject
configurations that may still satisfy a positive minimum requirement,
while retaining that pruning for zero-weight minimum-only terms. Add
regression coverage for both objective directions, zero crossings,
canceled explicit values, and three-component combined modifiers.

Move Codpiece plan and cache state, familiar contribution analysis, late
modifier calculation, gem-count validation, and canonical gem recursion
from MaximizerSpeculation into a package-private CodpieceSpeculation
collaborator. Leave only calculation and equipment-search integration
points in the general speculation class.
