+++ src/net/sourceforge/kolmafia/maximizer/MaximizerSpeculation.java
@@ -1,20 +1,15 @@
 package net.sourceforge.kolmafia.maximizer;
 
-import java.util.ArrayList;
 import java.util.EnumMap;
-import java.util.EnumSet;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.function.BiConsumer;
-import java.util.function.Supplier;
 import net.sourceforge.kolmafia.AdventureResult;
 import net.sourceforge.kolmafia.FamiliarData;
 import net.sourceforge.kolmafia.KoLCharacter;
 import net.sourceforge.kolmafia.KoLConstants;
 import net.sourceforge.kolmafia.KoLConstants.WeaponType;
 import net.sourceforge.kolmafia.KoLmafia;
-import net.sourceforge.kolmafia.ModifierType;
 import net.sourceforge.kolmafia.Modifiers;
 import net.sourceforge.kolmafia.RequestLogger;
 import net.sourceforge.kolmafia.Speculation;
@@ -22,7 +17,6 @@ import net.sourceforge.kolmafia.equipment.Slot;
 import net.sourceforge.kolmafia.equipment.SlotSet;
 import net.sourceforge.kolmafia.modifiers.BitmapModifier;
 import net.sourceforge.kolmafia.modifiers.BooleanModifier;
-import net.sourceforge.kolmafia.modifiers.DoubleModifier;
 import net.sourceforge.kolmafia.modifiers.StringModifier;
 import net.sourceforge.kolmafia.objectpool.ItemPool;
 import net.sourceforge.kolmafia.objectpool.SkillPool;
@@ -36,8 +30,6 @@ import net.sourceforge.kolmafia.session.EquipmentManager;
 
 public class MaximizerSpeculation extends Speculation
     implements Comparable<MaximizerSpeculation>, Cloneable {
-  private static final Slot[] CODPIECE_SLOTS = SlotSet.CODPIECE_SLOTS.toArray(Slot[]::new);
-
   boolean scored = false;
   private boolean tiebreakered = false;
   private boolean exceeded;
@@ -48,80 +40,7 @@ public class MaximizerSpeculation extends Speculation
   public boolean failed = false;
   public CheckedItem attachment;
   private boolean foldables = false;
-  private CodpieceSearch codpieceSearch;
-  private final Map<Integer, Modifiers> codpieceGemModifiers = new HashMap<>();
-  private final Map<Integer, Boolean> safeLateCodpieceGems = new HashMap<>();
-  private CodpiecePlan codpiecePlan;
-  private CodpiecePlan prioritizedCodpiecePlan;
-
-  private record CodpiecePlan(
-      List<CheckedItem> gems, Modifiers[] gemModifiers, boolean[] familiarDependentGems) {}
-
-  /** Reuses the expensive adjustment prefix while the search changes only late-safe gem effects. */
-  private static final class LateCodpieceCache {
-    private final Modifiers baseline;
-    private final Modifiers fightMods;
-    private final List<Slot> slots;
-    private final Modifiers[] gemModifiers;
-    private final boolean[] familiarDependentGems;
-    private final CodpiecePruning.FamiliarScoreContributions familiarScoreContributions;
-    private final Modifiers[] slotModifiers;
-    private final int[] slotGemIndexes;
-    private final Map<List<Integer>, KoLCharacter.AdjustmentPrefix> prefixes;
-    private int familiarSlots;
-
-    private LateCodpieceCache(
-        Modifiers baseline,
-        Modifiers fightMods,
-        List<Slot> slots,
-        Modifiers[] gemModifiers,
-        boolean[] familiarDependentGems,
-        CodpiecePruning.FamiliarScoreContributions familiarScoreContributions,
-        Map<List<Integer>, KoLCharacter.AdjustmentPrefix> prefixes) {
-      this.baseline = baseline;
-      this.fightMods = fightMods;
-      this.slots = slots;
-      this.gemModifiers = gemModifiers;
-      this.familiarDependentGems = familiarDependentGems;
-      this.familiarScoreContributions = familiarScoreContributions;
-      this.slotModifiers = new Modifiers[slots.size()];
-      this.slotGemIndexes = new int[slots.size()];
-      this.prefixes = prefixes;
-    }
-
-    private void select(int slotIndex, int gemIndex) {
-      this.slotModifiers[slotIndex] = this.gemModifiers[gemIndex];
-      this.slotGemIndexes[slotIndex] = gemIndex + 1;
-      if (this.familiarDependentGems[gemIndex]) {
-        this.familiarSlots++;
-      }
-    }
-
-    private void deselect(int slotIndex) {
-      int gemIndex = this.slotGemIndexes[slotIndex] - 1;
-      if (gemIndex >= 0 && this.familiarDependentGems[gemIndex]) {
-        this.familiarSlots--;
-      }
-      this.slotModifiers[slotIndex] = null;
-      this.slotGemIndexes[slotIndex] = 0;
-    }
-
-    private KoLCharacter.AdjustmentPrefix getPrefix(
-        Supplier<KoLCharacter.AdjustmentPrefix> factory) {
-      List<Integer> familiarGemIndexes = new ArrayList<>();
-      for (int encodedGemIndex : this.slotGemIndexes) {
-        if (encodedGemIndex == 0) {
-          continue;
-        }
-        int gemIndex = encodedGemIndex - 1;
-        if (this.familiarDependentGems[gemIndex]) {
-          familiarGemIndexes.add(gemIndex);
-        }
-      }
-      return this.prefixes.computeIfAbsent(
-          List.copyOf(familiarGemIndexes), ignored -> factory.get());
-    }
-  }
+  private CodpieceSpeculation codpiece = new CodpieceSpeculation(this);
 
   @Override
   public MaximizerSpeculation clone() {
@@ -132,7 +51,8 @@ public class MaximizerSpeculation extends Speculation
       if (this.mods != null) {
         copy.mods = new Modifiers(this.mods);
       }
-      copy.codpieceSearch = null;
+      // A clone is a frozen candidate: it must not retain an in-progress Codpiece search.
+      copy.codpiece = new CodpieceSpeculation(copy);
       return copy;
     } catch (CloneNotSupportedException e) {
       return null;
@@ -141,37 +61,13 @@ public class MaximizerSpeculation extends Speculation
 
   @Override
   public Modifiers calculate() {
-    LateCodpieceCache cache = this.codpieceSearch == null ? null : this.codpieceSearch.cache;
-    if (cache == null) {
+    var result = this.codpiece.calculateModifiers();
+    if (result == null) {
       return super.calculate();
-    }
-
-    Modifiers newModifiers;
-    Modifiers fightMods;
-    if (cache.familiarSlots == 0) {
-      newModifiers = new Modifiers(cache.baseline);
-      for (Modifiers gemModifiers : cache.slotModifiers) {
-        newModifiers.add(gemModifiers);
-      }
-      fightMods = cache.fightMods;
-    } else {
-      // Familiar-dependent gems must be present when familiar effects are calculated.
-      var prefix = cache.getPrefix(this::primeFamiliarCodpiecePrefix);
-      newModifiers = new Modifiers(prefix.modifiers());
-      for (int encodedGemIndex : cache.slotGemIndexes) {
-        if (encodedGemIndex == 0) {
-          continue;
-        }
-        int gemIndex = encodedGemIndex - 1;
-        if (!cache.familiarDependentGems[gemIndex]) {
-          newModifiers.add(cache.gemModifiers[gemIndex]);
-        }
-      }
-      fightMods = prefix.fightMods();
     }
     this.mods =
         KoLCharacter.applyAdjustmentSuffix(
-            false, newModifiers, fightMods, this.equipment, this.getEffects(), true);
+            false, result.modifiers(), result.fightMods(), this.equipment, this.getEffects(), true);
     this.calculated = true;
     return this.mods;
   }
@@ -803,364 +699,11 @@ public class MaximizerSpeculation extends Speculation
       this.equipment.put(Slot.OFFHAND, EquipmentRequest.UNEQUIP);
     }
 
-    boolean wearingCodpiece =
-        this.equipment.values().stream()
-            .anyMatch(item -> item != null && item.getItemId() == ItemPool.THE_ETERNITY_CODPIECE);
-    if (!wearingCodpiece) {
-      this.releaseCodpieceGemsNeededElsewhere();
-      this.checkBest();
-      this.restore(mark);
-      return;
-    }
-
-    List<Slot> codpieceSlots =
-        SlotSet.CODPIECE_SLOTS.stream().filter(Maximizer.eval::slotEnabled).toList();
-    for (Slot slot : codpieceSlots) {
-      this.equipment.put(slot, EquipmentRequest.UNEQUIP);
-    }
-    if (!this.hasEnoughCodpieceGems()) {
-      this.restore(mark);
-      return;
-    }
-
-    CodpiecePlan plan = this.getCodpiecePlan(possibles.get(Slot.CODPIECE1));
-    List<CheckedItem> codpieceGems = plan.gems();
-    try {
-      // Saturation does not model tiebreaks, beeosity, or removal of initially equipped gems.
-      boolean canCollapseSaturatedScore =
-          !Maximizer.eval.isUsingTiebreaker()
-              && !KoLCharacter.inBeecore()
-              && codpieceSlots.stream()
-                  .allMatch(
-                      slot -> EquipmentManager.getEquipment(slot).equals(EquipmentRequest.UNEQUIP));
-      LateCodpieceCache cache =
-          this.canUseLateCodpieceCache(codpieceGems, codpieceSlots)
-              ? this.primeLateCodpieceCache(plan, codpieceSlots)
-              : null;
-      CodpieceSearch search =
-          new CodpieceSearch(codpieceGems, codpieceSlots, cache, canCollapseSaturatedScore);
-      this.codpieceSearch = search;
-      this.calculated = false;
-      if (Maximizer.eval.isUsingTiebreaker()
-          && Maximizer.eval.areScoreModifiersSaturated(this.calculate())) {
-        if (this.prioritizedCodpiecePlan == null) {
-          this.prioritizedCodpiecePlan =
-              this.createCodpiecePlan(Maximizer.eval.prioritizeCodpieceGems(codpieceGems));
-        }
-        codpieceGems = this.prioritizedCodpiecePlan.gems();
-        cache = this.primeLateCodpieceCache(this.prioritizedCodpiecePlan, codpieceSlots);
-        search = new CodpieceSearch(codpieceGems, codpieceSlots, cache, false);
-      }
-      this.codpieceSearch = search;
-      this.codpieceSearch.run();
-    } finally {
-      this.codpieceSearch = null;
-      this.restore(mark);
-    }
+    this.codpiece.trySlots(mark, possibles.get(Slot.CODPIECE1));
   }
 
-  private void releaseCodpieceGemsNeededElsewhere() {
-    for (Slot slot : CODPIECE_SLOTS) {
-      if (!Maximizer.eval.slotEnabled(slot)) {
-        continue;
-      }
-
-      AdventureResult gem = this.equipment.get(slot);
-      if (gem == null || gem.equals(EquipmentRequest.UNEQUIP)) {
-        continue;
-      }
-
-      CheckedItem equippedElsewhere =
-          this.equipment.entrySet().stream()
-              .filter(entry -> !SlotSet.CODPIECE_SLOTS.contains(entry.getKey()))
-              .map(Map.Entry::getValue)
-              .filter(gem::equals)
-              .filter(CheckedItem.class::isInstance)
-              .map(CheckedItem.class::cast)
-              .findFirst()
-              .orElse(null);
-      if (equippedElsewhere == null) {
-        continue;
-      }
-
-      long used = this.equipment.values().stream().filter(gem::equals).count();
-      if (used > equippedElsewhere.getAvailableCount()) {
-        this.equipment.put(slot, EquipmentRequest.UNEQUIP);
-      }
-    }
-  }
-
-  private boolean canUseLateCodpieceCache(List<CheckedItem> possibles, List<Slot> slots) {
-    for (Slot slot : slots) {
-      AdventureResult equipped = this.equipment.get(slot);
-      if (equipped != null
-          && equipped != EquipmentRequest.UNEQUIP
-          && !this.isSafeLateCodpieceGem(equipped.getItemId())) {
-        return false;
-      }
-    }
-
-    for (CheckedItem possible : possibles) {
-      if (!this.isSafeLateCodpieceGem(possible.getItemId())) {
-        return false;
-      }
-    }
-
-    return true;
-  }
-
-  private CodpiecePlan getCodpiecePlan(List<CheckedItem> possibles) {
-    if (this.codpiecePlan == null) {
-      this.codpiecePlan =
-          this.createCodpiecePlan(
-              possibles.stream()
-                  .filter(
-                      gem -> gem.getCount() > 0 && EquipmentRequest.isCodpieceGem(gem.getItemId()))
-                  .toList());
-    }
-    return this.codpiecePlan;
-  }
-
-  private CodpiecePlan createCodpiecePlan(List<CheckedItem> gems) {
-    Modifiers[] gemModifiers = new Modifiers[gems.size()];
-    boolean[] familiarDependentGems = new boolean[gems.size()];
-    for (int i = 0; i < gems.size(); i++) {
-      gemModifiers[i] = this.getCodpieceGemModifiers(gems.get(i).getItemId());
-      familiarDependentGems[i] = CodpiecePruning.affectsFamiliarCalculation(gemModifiers[i]);
-    }
-    return new CodpiecePlan(gems, gemModifiers, familiarDependentGems);
-  }
-
-  private LateCodpieceCache primeLateCodpieceCache(CodpiecePlan plan, List<Slot> slots) {
-    var mark = this.mark();
-    try {
-      for (Slot slot : slots) {
-        this.equipment.put(slot, EquipmentRequest.UNEQUIP);
-      }
-
-      var prefix = this.recalculateCodpiecePrefix(this.equipment);
-      Map<List<Integer>, KoLCharacter.AdjustmentPrefix> prefixes = new HashMap<>();
-      prefixes.put(List.of(), prefix);
-      CodpiecePruning.FamiliarScoreContributions familiarScoreContributions =
-          this.findFamiliarScoreContributions(
-              plan.gems(),
-              slots,
-              prefix.modifiers(),
-              prefix.familiarWeightInputs(),
-              plan.gemModifiers(),
-              plan.familiarDependentGems());
-      return new LateCodpieceCache(
-          prefix.modifiers(),
-          prefix.fightMods(),
-          slots,
-          plan.gemModifiers(),
-          plan.familiarDependentGems(),
-          familiarScoreContributions,
-          prefixes);
-    } finally {
-      this.restore(mark);
-    }
-  }
-
-  private CodpiecePruning.FamiliarScoreContributions findFamiliarScoreContributions(
-      List<CheckedItem> possibles,
-      List<Slot> slots,
-      Modifiers baseline,
-      Modifiers.FamiliarWeightInputs familiarWeightInputs,
-      Modifiers[] gemModifiers,
-      boolean[] familiarDependentGems) {
-    EnumSet<DoubleModifier> scored = Maximizer.eval.familiarDependentScoreModifiers();
-    if (scored.isEmpty()) {
-      return new CodpiecePruning.FamiliarScoreContributions(-1, Map.of());
-    }
-    int familiarGemIndex = -1;
-    for (int i = 0; i < familiarDependentGems.length; i++) {
-      if (!familiarDependentGems[i]) {
-        continue;
-      }
-      if (familiarGemIndex != -1) {
-        return null;
-      }
-      familiarGemIndex = i;
-    }
-    if (familiarGemIndex == -1) {
-      return new CodpiecePruning.FamiliarScoreContributions(-1, Map.of());
-    }
-
-    Map<DoubleModifier, CodpiecePruning.ContributionRange> ranges =
-        new EnumMap<>(DoubleModifier.class);
-    Map<DoubleModifier, CodpiecePruning.ContributionRange> baselines =
-        new EnumMap<>(DoubleModifier.class);
-    CheckedItem gem = possibles.get(familiarGemIndex);
-    double weightAdjustment =
-        CodpiecePruning.familiarWeightAdjustment(gemModifiers[familiarGemIndex]);
-    if (!Double.isFinite(weightAdjustment)) {
-      return null;
-    }
-    int copies = Math.min(slots.size(), gem.getCount() - (int) this.countEquipped(gem.getItemId()));
-    var familiar = this.getFamiliar();
-    var baselineEffect = baseline.familiarEffect(familiar, familiarWeightInputs, 0);
-    var previous = baselineEffect;
-    double[] directExperienceDeltas = new double[copies];
-    double[] generalExperienceDeltas = new double[copies];
-    for (int copy = 1; copy <= copies; copy++) {
-      var current =
-          baseline.familiarEffect(familiar, familiarWeightInputs, copy * weightAdjustment);
-      directExperienceDeltas[copy - 1] =
-          current.primeStatExperience() - previous.primeStatExperience();
-      generalExperienceDeltas[copy - 1] =
-          current.generalExperience() - previous.generalExperience();
-      for (DoubleModifier modifier : scored) {
-        double contribution =
-            switch (modifier) {
-              case ITEMDROP -> current.itemDrop() - previous.itemDrop();
-              case MEATDROP -> current.meatDrop() - previous.meatDrop();
-              case EXPERIENCE, MUS_EXPERIENCE, MYS_EXPERIENCE, MOX_EXPERIENCE -> 0.0;
-              default -> 0.0;
-            };
-        if (contribution == 0.0) {
-          continue;
-        }
-        ranges.merge(
-            modifier,
-            new CodpiecePruning.ContributionRange(contribution, contribution),
-            (range, ignored) -> range.include(contribution));
-      }
-      previous = current;
-    }
-    for (DoubleModifier modifier : scored) {
-      if (!CodpiecePruning.isExperienceScoreModifier(modifier)) {
-        continue;
-      }
-      baselines.put(
-          modifier,
-          this.findFamiliarExperienceRange(
-              modifier,
-              possibles,
-              slots.size(),
-              baseline,
-              gemModifiers,
-              familiarGemIndex,
-              new double[] {0.0},
-              new double[] {baselineEffect.generalExperience()}));
-      ranges.put(
-          modifier,
-          this.findFamiliarExperienceRange(
-              modifier,
-              possibles,
-              slots.size(),
-              baseline,
-              gemModifiers,
-              familiarGemIndex,
-              directExperienceDeltas,
-              generalExperienceDeltas));
-    }
-    return new CodpiecePruning.FamiliarScoreContributions(
-        familiarGemIndex, Map.copyOf(ranges), Map.copyOf(baselines));
-  }
-
-  private CodpiecePruning.ContributionRange findFamiliarExperienceRange(
-      DoubleModifier scoreModifier,
-      List<CheckedItem> possibles,
-      int slotCount,
-      Modifiers baseline,
-      Modifiers[] gemModifiers,
-      int familiarGemIndex,
-      double[] directDeltas,
-      double[] generalDeltas) {
-    int[] remaining = new int[possibles.size()];
-    var relevant = new ArrayList<Integer>();
-    for (int i = 0; i < possibles.size(); i++) {
-      remaining[i] =
-          Math.min(
-              slotCount,
-              possibles.get(i).getCount() - (int) this.countEquipped(possibles.get(i).getItemId()));
-      if (i != familiarGemIndex
-          && CodpiecePruning.affectsExperience(gemModifiers[i], scoreModifier)) {
-        relevant.add(i);
-      }
-    }
-    var range =
-        new CodpiecePruning.ContributionRange(Double.POSITIVE_INFINITY, Double.NEGATIVE_INFINITY);
-    return collectFamiliarExperienceRange(
-        scoreModifier,
-        baseline,
-        gemModifiers,
-        remaining,
-        relevant,
-        new int[possibles.size()],
-        directDeltas,
-        generalDeltas,
-        range,
-        0,
-        0,
-        slotCount);
-  }
-
-  private static CodpiecePruning.ContributionRange collectFamiliarExperienceRange(
-      DoubleModifier scoreModifier,
-      Modifiers current,
-      Modifiers[] gemModifiers,
-      int[] remaining,
-      List<Integer> relevant,
-      int[] used,
-      double[] directDeltas,
-      double[] generalDeltas,
-      CodpiecePruning.ContributionRange range,
-      int start,
-      int selected,
-      int slotCount) {
-    for (int copy = 0; copy < directDeltas.length; copy++) {
-      range =
-          range.include(
-              CodpiecePruning.familiarExperienceContribution(
-                  scoreModifier, current, directDeltas[copy], generalDeltas[copy]));
-    }
-    if (selected == slotCount) {
-      return range;
-    }
-    for (int relevantIndex = start; relevantIndex < relevant.size(); relevantIndex++) {
-      int gemIndex = relevant.get(relevantIndex);
-      if (used[gemIndex] >= remaining[gemIndex]) {
-        continue;
-      }
-      var next = new Modifiers(current);
-      CodpiecePruning.addExperienceInputs(next, gemModifiers[gemIndex]);
-      used[gemIndex]++;
-      range =
-          collectFamiliarExperienceRange(
-              scoreModifier,
-              next,
-              gemModifiers,
-              remaining,
-              relevant,
-              used,
-              directDeltas,
-              generalDeltas,
-              range,
-              relevantIndex,
-              selected + 1,
-              slotCount);
-      used[gemIndex]--;
-    }
-    return range;
-  }
-
-  private KoLCharacter.AdjustmentPrefix primeFamiliarCodpiecePrefix() {
-    LateCodpieceCache cache = this.codpieceSearch.cache;
-    Map<Slot, AdventureResult> equipment = new EnumMap<>(this.equipment);
-    for (int slotIndex = 0; slotIndex < cache.slots.size(); slotIndex++) {
-      int encodedGemIndex = cache.slotGemIndexes[slotIndex];
-      if (encodedGemIndex == 0 || !cache.familiarDependentGems[encodedGemIndex - 1]) {
-        equipment.put(cache.slots.get(slotIndex), EquipmentRequest.UNEQUIP);
-      }
-    }
-
-    return this.recalculateCodpiecePrefix(equipment);
-  }
-
-  private KoLCharacter.AdjustmentPrefix recalculateCodpiecePrefix(
-      Map<Slot, AdventureResult> equipment) {
+  /** Used by {@link CodpieceSpeculation} to build the late-adjustment prefix it caches. */
+  KoLCharacter.AdjustmentPrefix recalculateCodpiecePrefix(Map<Slot, AdventureResult> equipment) {
     return KoLCharacter.recalculateAdjustmentsPrefix(
         false,
         this.getMindControlLevel(),
@@ -1174,218 +717,6 @@ public class MaximizerSpeculation extends Speculation
         this.getBoomBox(),
         this.getModeables(),
         true);
-  }
-
-  private boolean isSafeLateCodpieceGem(int itemId) {
-    return this.safeLateCodpieceGems.computeIfAbsent(
-        itemId,
-        id ->
-            CodpiecePruning.hasOnlySupportedLateCalculationModifiers(
-                this.getCodpieceGemModifiers(id)));
-  }
-
-  private Modifiers getCodpieceGemModifiers(int itemId) {
-    return this.codpieceGemModifiers.computeIfAbsent(
-        itemId, id -> ModifierDatabase.getModifiers(ModifierType.ETERNITY_CODPIECE, id));
-  }
-
-  private int countEquipmentWith(BooleanModifier modifier) {
-    int count = 0;
-    for (AdventureResult item : this.equipment.values()) {
-      if (item == null) {
-        continue;
-      }
-      Modifiers modifiers = ModifierDatabase.getItemModifiers(item.getItemId());
-      if (modifiers != null && modifiers.getBoolean(modifier)) {
-        count++;
-      }
-    }
-    return count;
-  }
-
-  /** Enumerates canonical gem multisets and rejects branches using conservative score ceilings. */
-  private final class CodpieceSearch {
-    private final List<CheckedItem> gems;
-    private final List<Slot> slots;
-    private final int[] remaining;
-    private final int[] initialRemaining;
-    private final boolean[] required;
-    private final int requiredCount;
-    private final LateCodpieceCache cache;
-    private final boolean canCollapseSaturatedScore;
-    private CodpiecePruning.BranchBounds bounds;
-    private boolean tiebreakBoundInitialized;
-
-    private CodpieceSearch(
-        List<CheckedItem> gems,
-        List<Slot> slots,
-        LateCodpieceCache cache,
-        boolean canCollapseSaturatedScore) {
-      this.gems = gems;
-      this.slots = slots;
-      this.remaining = new int[gems.size()];
-      this.required = new boolean[gems.size()];
-      this.cache = cache;
-      this.canCollapseSaturatedScore = canCollapseSaturatedScore;
-
-      int requiredCount = 0;
-      for (int i = 0; i < gems.size(); i++) {
-        CheckedItem gem = gems.get(i);
-        long used = MaximizerSpeculation.this.countEquipped(gem.getItemId());
-        this.remaining[i] = gem.getCount() - (int) used;
-        if (gem.requiredFlag && used == 0) {
-          this.required[i] = true;
-          requiredCount++;
-        }
-      }
-      this.initialRemaining = this.remaining.clone();
-      this.requiredCount = requiredCount;
-    }
-
-    private void run() throws MaximizerInterruptedException {
-      CodpieceScoreBound scoreUpperBound = this.createScoreUpperBound();
-      this.bounds =
-          new CodpiecePruning.BranchBounds(
-              scoreUpperBound,
-              null,
-              new CodpiecePruning.BooleanUpperBound(
-                  this.gems, this.remaining, this.slots.size(), BooleanModifier.DROPS_ITEMS),
-              new CodpiecePruning.BooleanUpperBound(
-                  this.gems, this.remaining, this.slots.size(), BooleanModifier.DROPS_MEAT));
-      this.search(0, 0, this.requiredCount);
-    }
-
-    private CodpieceScoreBound createScoreUpperBound() {
-      return this.cache == null
-          ? null
-          : Maximizer.eval.createTheoreticalCodpieceScoreUpperBound(
-              this.cache.baseline,
-              this.cache.gemModifiers,
-              this.remaining,
-              this.slots.size(),
-              MaximizerSpeculation.this.equipment,
-              MaximizerSpeculation.this.getModeables(),
-              this.gems,
-              this.cache.familiarScoreContributions);
-    }
-
-    private void search(int start, int slotIndex, int requiredCount)
-        throws MaximizerInterruptedException {
-      boolean scoreSaturated = false;
-      boolean canMeetRequirements = true;
-      if (this.bounds.score() != null) {
-        int remainingSlots = this.slots.size() - slotIndex;
-        double upperScore = this.bounds.score().estimate(start, this.remaining, remainingSlots);
-        canMeetRequirements &=
-            this.bounds.score().canMeetMinimum(start, this.remaining, remainingSlots, upperScore);
-        if (!Maximizer.best.failed || !canMeetRequirements) {
-          double bestScore = Maximizer.best.getScore();
-          if (upperScore < bestScore) {
-            return;
-          }
-          if (Double.compare(upperScore, bestScore) == 0 && !KoLCharacter.inBeecore()) {
-            // Keep this tie pruning in compareTo order: item drops, meat drops, then tiebreak
-            // score.
-            CodpieceScoreBound tiebreakUpperBound = this.getTiebreakUpperBound(slotIndex);
-            int bestItemDroppers = Maximizer.best.countEquipmentWith(BooleanModifier.DROPS_ITEMS);
-            int itemDropperCeiling =
-                MaximizerSpeculation.this.countEquipmentWith(BooleanModifier.DROPS_ITEMS)
-                    + this.bounds
-                        .itemDroppers()
-                        .estimateAdditional(start, this.remaining, remainingSlots);
-            if (itemDropperCeiling < bestItemDroppers) {
-              return;
-            }
-            if (itemDropperCeiling == bestItemDroppers && tiebreakUpperBound != null) {
-              int bestMeatDroppers = Maximizer.best.countEquipmentWith(BooleanModifier.DROPS_MEAT);
-              int meatDropperCeiling =
-                  MaximizerSpeculation.this.countEquipmentWith(BooleanModifier.DROPS_MEAT)
-                      + this.bounds
-                          .meatDroppers()
-                          .estimateAdditional(start, this.remaining, remainingSlots);
-              if (meatDropperCeiling < bestMeatDroppers
-                  || (meatDropperCeiling == bestMeatDroppers
-                      && tiebreakUpperBound.estimate(start, this.remaining, remainingSlots)
-                          < Maximizer.best.getTiebreaker())) {
-                return;
-              }
-            }
-          }
-        }
-        scoreSaturated = this.bounds.score().isScoreSaturated(start, this.remaining, upperScore);
-      }
-      if (requiredCount == 0) {
-        MaximizerSpeculation.this.checkBest(true);
-        if (this.canCollapseSaturatedScore
-            && scoreSaturated
-            && (!MaximizerSpeculation.this.failed || !canMeetRequirements)) {
-          return;
-        }
-      }
-      if (slotIndex == this.slots.size() || requiredCount > this.slots.size() - slotIndex) {
-        return;
-      }
-
-      int firstRequired = -1;
-      for (int i = start; i < this.required.length; i++) {
-        if (this.required[i]) {
-          firstRequired = i;
-          break;
-        }
-      }
-
-      Slot slot = this.slots.get(slotIndex);
-      for (int i = start; i < this.gems.size(); i++) {
-        if (firstRequired != -1 && i > firstRequired) {
-          break;
-        }
-        if (this.remaining[i] == 0) {
-          continue;
-        }
-
-        boolean satisfiesRequirement = this.required[i];
-        this.remaining[i]--;
-        this.required[i] = false;
-        MaximizerSpeculation.this.equipment.put(slot, this.gems.get(i));
-        if (this.cache != null) {
-          this.cache.select(slotIndex, i);
-        }
-        this.bounds.select(i);
-        this.search(i, slotIndex + 1, requiredCount - (satisfiesRequirement ? 1 : 0));
-        this.bounds.deselect(i);
-        MaximizerSpeculation.this.equipment.put(slot, EquipmentRequest.UNEQUIP);
-        if (this.cache != null) {
-          this.cache.deselect(slotIndex);
-        }
-        this.required[i] = satisfiesRequirement;
-        this.remaining[i]++;
-      }
-    }
-
-    private CodpieceScoreBound getTiebreakUpperBound(int selectedCount) {
-      if (this.tiebreakBoundInitialized) {
-        return this.bounds.tiebreaker();
-      }
-
-      this.tiebreakBoundInitialized = true;
-      CodpieceScoreBound tiebreakUpperBound =
-          Maximizer.eval.createTheoreticalCodpieceTiebreakerUpperBound(
-              this.cache.baseline,
-              this.cache.gemModifiers,
-              this.initialRemaining,
-              this.slots.size(),
-              this.cache.familiarScoreContributions);
-      for (int i = 0; tiebreakUpperBound != null && i < selectedCount; i++) {
-        tiebreakUpperBound.select(this.cache.slotGemIndexes[i] - 1);
-      }
-      this.bounds =
-          new CodpiecePruning.BranchBounds(
-              this.bounds.score(),
-              tiebreakUpperBound,
-              this.bounds.itemDroppers(),
-              this.bounds.meatDroppers());
-      return tiebreakUpperBound;
-    }
   }
 
   /** Applies each candidate to its own clone of baseline and returns the best-scoring one. */
@@ -1405,12 +736,12 @@ public class MaximizerSpeculation extends Speculation
     return best;
   }
 
-  private void checkBest() throws MaximizerInterruptedException {
+  void checkBest() throws MaximizerInterruptedException {
     this.checkBest(false);
   }
 
-  private void checkBest(boolean codpieceCountsValid) throws MaximizerInterruptedException {
-    if (!codpieceCountsValid && !this.hasEnoughCodpieceGems()) {
+  void checkBest(boolean codpieceCountsValid) throws MaximizerInterruptedException {
+    if (!codpieceCountsValid && !this.codpiece.hasEnoughCodpieceGems()) {
       return;
     }
 
@@ -1443,44 +774,6 @@ public class MaximizerSpeculation extends Speculation
     if (Maximizer.combinationLimit != 0 && Maximizer.bestChecked >= Maximizer.combinationLimit) {
       throw new MaximizerLimitException();
     }
-  }
-
-  private long countEquipped(int itemId) {
-    long count = 0;
-    for (AdventureResult item : this.equipment.values()) {
-      if (item != null && item.getItemId() == itemId) {
-        count++;
-      }
-    }
-    return count;
-  }
-
-  private boolean hasEnoughCodpieceGems() {
-    boolean hasSlottedGem = false;
-    for (Slot slot : CODPIECE_SLOTS) {
-      AdventureResult item = this.equipment.get(slot);
-      if (item != null && !item.equals(EquipmentRequest.UNEQUIP)) {
-        hasSlottedGem = true;
-        break;
-      }
-    }
-    if (!hasSlottedGem) {
-      return true;
-    }
-
-    for (AdventureResult item : this.equipment.values()) {
-      if (!(item instanceof CheckedItem checked)
-          || !EquipmentRequest.isCodpieceGem(item.getItemId())) {
-        continue;
-      }
-
-      int itemId = item.getItemId();
-      int used = (int) this.countEquipped(itemId);
-      if (used > checked.getAvailableCount()) {
-        return false;
-      }
-    }
-    return true;
   }
 
   private static int getMutex(AdventureResult item) {
